+++ db/app/msg/src/full/R__09.PACKAGE_SPEC.HTTP_REQUEST_RESPONSE_PKG.sql
@@ -28,7 +28,8 @@ function get_cookie_idx
 , p_name in varchar2 -- The cookie name to search for
 , p_ignore_case in boolean default false -- Must we ignore case (default NO)?
 )
-return positive; -- null (not found) or a positive number (found)
+return positive -- null (not found) or a positive number (found)
+deterministic;
 /**
 
 > Cookie names are case-sensitive.
@@ -39,19 +40,37 @@ See also [Is the name of a cookie case sensitive?](https://stackoverflow.com/que
 
 **/
 
+function get_cookie
+( p_cookies in http_cookie_tab_typ -- The cookies
+, p_name in varchar2 -- The cookie name to search for
+, p_ignore_case in boolean default false -- Must we ignore case (default NO)?
+)
+return varchar2 -- null (not found) or the cookie (found)
+deterministic;
+/** Uses get_cookie_idx to return the value when found. **/
+
 function get_property_idx
 ( p_properties in property_tab_typ -- The properties (for instance HTTP headers)
 , p_name in varchar2 -- The property to search for
 , p_ignore_case in boolean default true -- Must we ignore case (default YES)?
 )
-return positive; -- null (not found) or a positive number (found)
-
+return positive -- null (not found) or a positive number (found)
+deterministic;
 /**
 
 > HTTP header names are case-insensitive (according to RFC 2616).
 
 See also [Are HTTP headers case-sensitive?](https://stackoverflow.com/questions/5258977/are-http-headers-case-sensitive).
 **/
+
+function get_property
+( p_properties in property_tab_typ -- The properties (for instance HTTP headers)
+, p_name in varchar2 -- The property to search for
+, p_ignore_case in boolean default true -- Must we ignore case (default YES)?
+)
+return varchar2 -- null (not found) or the property (found)
+deterministic;
+/** Uses get_property_idx to return the value when found. **/
 
 function get_http_status_descr
 ( p_http_status_code in positiven -- Should be > 0