+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Types of changes:
 - [The message subsystem needs to handle HTTP responses with respect to HTTP rate limits.](https://github.com/paulissoft/oracle-tools/issues/211)
 - [Base the smart pool data source directly on the Hikari and Oracle UCP pool data sources.](https://github.com/paulissoft/oracle-tools/issues/212)
 - [Some unit tests for web_service_pkg fail.](https://github.com/paulissoft/oracle-tools/issues/213)
+- [The API_TIME_PKG must also support milliseconds since the Unix epoch, i.e. 1-JAN-1970 UTC.](https://github.com/paulissoft/oracle-tools/issues/214)
 
 ## [4.4.2] - 2025-04-17